backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
hdfs==2.7.3
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
celery==5.3.4
httpx==0.25.1
aiofiles==23.2.1
alembic==1.12.1
click==8.1.7
psutil==5.9.5